Split up console tab component into a separate file
This component is too large now to not have its own file
