+++ src/vs/workbench/contrib/positronConsole/browser/components/consoleTab.css
@@ -1,0 +1,185 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (C) 2025-2026 Posit Software, PBC. All rights reserved.
+ *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+.tabs-container .tab-button {
+	background-color: inherit;
+	/*
+	 * The transparent border allows for a smooth transition when switching tabs.
+	 * The active tab will have a solid border, while the inactive tabs will have
+	 * a transparent border. The transparent border prevents flickering when
+	 * switching between tabs by maintaining the same width for all tabs.
+	 */
+	border-left: 1px solid transparent;
+	color: var(--vscode-foreground);
+	cursor: pointer;
+	padding: 4px 10px;
+	width: 100%;
+	text-align: left;
+	white-space: nowrap;
+	overflow: hidden;
+	text-overflow: ellipsis;
+
+	/* Use this property to take padding into account when sizing (prevents overflow) */
+	box-sizing: border-box;
+	display: flex;
+	flex-direction: column;
+}
+
+.tabs-container .tab-button .tab-header {
+	display: flex;
+	align-items: center;
+	height: 22px;
+	width: 100%;
+}
+
+.tabs-container .tab-button:hover {
+	background-color: var(--vscode-list-hoverBackground);
+}
+
+.tabs-container .tab-button--active {
+	background-color: var(--vscode-list-inactiveSelectionBackground);
+	border-left: 1px solid var(--vscode-panelTitle-activeBorder);
+}
+
+.tabs-container .tab-button--active:hover {
+	background-color: var(--vscode-list-activeSelectionBackground);
+}
+
+.tabs-container .tab-button .icon {
+	height: 15px;
+	width: 15px;
+	margin: 0 6px
+}
+
+.tabs-container .tab-button .session-name {
+	/* Grow items after session name to fill available space */
+	flex: 1;
+
+	margin: 0;
+	line-height: 22px;
+	overflow: hidden;
+	min-width: 0;
+	text-overflow: ellipsis;
+}
+
+.tabs-container .tab-button .session-name-input {
+	background-color: var(--vscode-input-background);
+	border: 1px solid var(--vscode-input-border);
+	color: var(--vscode-input-foreground);
+
+	/* Grow input to fill available space */
+	flex: 1;
+}
+
+.tabs-container .tab-button .session-name-input:focus {
+	border-color: var(--vscode-focusBorder);
+}
+
+.tabs-container .tab-button .delete-button {
+	/* delete button is hidden by default */
+	display: none;
+
+	background-color: transparent;
+	border: none;
+	cursor: pointer;
+	height: 22px;
+	padding: 2px;
+
+	border-radius: 5px;
+}
+
+.tabs-container .tab-button:hover .delete-button,
+.tabs-container .tab-button:focus-within .delete-button {
+	/* Show delete button on hover or when tab has focus */
+	display: unset;
+}
+
+.tabs-container .tab-button .delete-button:hover {
+	background-color: var(--vscode-toolbar-hoverBackground);
+}
+
+.tabs-container .tab-button .delete-button:focus {
+	outline: 1px solid var(--vscode-focusBorder);
+}
+
+.tabs-container .tab-button .delete-button .codicon {
+	height: 15px;
+	width: 15px;
+	vertical-align: text-bottom;
+}
+
+/* Resource usage section */
+.tabs-container .tab-button .resource-usage-section {
+	width: 100%;
+	margin-top: 4px;
+}
+
+/* Resource usage graph */
+.tabs-container .tab-button .resource-usage-graph {
+	display: block;
+}
+
+.tabs-container .tab-button .resource-usage-line {
+	fill: none;
+	stroke: var(--vscode-charts-lines, var(--vscode-foreground));
+	stroke-width: 1.5;
+}
+
+.tabs-container .tab-button .resource-usage-fill {
+	fill: var(--vscode-charts-lines, var(--vscode-foreground));
+	opacity: 0.2;
+}
+
+/* Resource usage section - enable container queries */
+.tabs-container .tab-button .resource-usage-section {
+	container-type: inline-size;
+}
+
+/* Resource usage stats */
+.tabs-container .tab-button .resource-usage-stats {
+	display: flex;
+	justify-content: space-between;
+	font-size: 9px;
+	text-transform: uppercase;
+	letter-spacing: 0.5px;
+	color: var(--vscode-descriptionForeground);
+	margin-top: 2px;
+	margin-bottom: 0px;
+}
+
+.tabs-container .tab-button .resource-usage-cpu,
+.tabs-container .tab-button .resource-usage-memory {
+	white-space: nowrap;
+	display: flex;
+	gap: 4px;
+	flex-direction: row;
+}
+
+
+.tabs-container .tab-button .resource-usage-value {
+	margin: 0px;
+}
+
+/* Hide labels when container is too narrow */
+@container (max-width: 115px) {
+	.resource-usage-stats .resource-usage-label {
+		display: none;
+	}
+}
+
+/*
+ * Hide CPU usage when it's even narrower (at its minimum width).
+ *
+ * The 64px value corresponds to the MINIMUM_CONSOLE_TAB_LIST_WIDTH, and should be
+ * kept in sync with that constant.
+ */
+@container (max-width: 64px) {
+	.tabs-container .tab-button .resource-usage-stats .resource-usage-cpu {
+		display: none;
+	}
+	.tabs-container .tab-button .resource-usage-stats {
+		justify-content: flex-end;
+	}
+}
